topicTracker 0.5!
Finished the preliminary build of topicTracker! This is a command line
utility which accepts user input of topic names and then tracks the time
spent on each subject. Also got a new piece of example code called
string_File_Output.c which describes the basics of string file output
(creative, I know). Adjusted filenames to be more searchable as well.